oda_project/requirements.txt
asgiref==3.5.0
black==22.1.0
click==8.0.3
Django==3.2.12
django-admin-interface==0.18.6
django-colorfield==0.6.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.4.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tomli==2.0.0
typing_extensions==4.0.1
